Flow analysis: replace context toString methods with a less fragile mechanism.
Adds getters `_debugFields` and `_debugType` to the `_FlowContext`
base class, and adds a single implementation of
`_FlowContext.toString` that builds a representation of the context
based on them.  This replaces the implementations of `toString` in all
the classes derived from `_FlowContext`, which were more difficult to
get right and keep synchronized with code changes.

Also changes `_TryFinallyContext` from a `late final` variable to a
nullable variable.  This sacrifices a tiny bit of safety, but has the
advantage of allowing `toString` to show the value if it's been
initialized, and avoid crashing if it hasn't.

Change-Id: I0ca3ddfed21934c54eaba912c84edda79c8eadfc
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/276202
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>