nnbd: fix late inference for cascades
Cascaded calls usually end up being `StatementExpression`s, and
for those dummy edges are added. But in a cascade, the result of
the expression is the same as the target, so these dummy edges
happen to affect nullability of cascade's target.

This fix just drops adding dummy edges for `StatementExpression`s
those expression is a cascade. I guess the better fix would be to
make dummy edges into suggestions, since AFAIUT, their meaning is
"it's fine to make this nullable" vs "this shouldn't be nullable"
added by the method invocation, so the latter should ideally win.
Not sure that will fit into the algorithm though.

Fixes: https://github.com/dart-lang/sdk/issues/49689
Tested: repro test now passes
Change-Id: I62b408c7307df4ea485b8a925cf4537fd9890024
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/277681
Reviewed-by: Paul Berry <paulberry@google.com>