Move `isSetupAssignment` to the base class
This field was added to track the assignments in `setUp` function
in tests, so allow better `late` inference for variables that get
assigned in `setUp`. But this might be needed not only for
`EspressionChecksOrigin` but also for `DynamicAssignmentOrigin` and
`AssignmentFromAngularInjectorGetOrigin`, so I moved this field to
the base `EdgeOrigin` class.

Tested: added a test that fails without this change.
Change-Id: I9c7c6b2e8b21f94e80e505cf24f2072bd8df703a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/277682
Reviewed-by: Paul Berry <paulberry@google.com>