[analyzer] Add facility to instantiate record types
This adds the `recordTypeWith` method to `TypeProvider`, a method
instantiating a record with provided positional and named fields.

By instantiating specific record types, we can match user-defined types
against specific record structures, for instance to validate that a
user-defined type is compatible to a certain record. For me, this would
be a helpful addition to the analyzer's public API.

Change-Id: I52e70eb0cbb0a0edf919b7b0f8901e6de1e00a64
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/277401
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>