[io] Improve scheduling of IO worker threads
The allocation of the at most 32 IO worker threads to IO requests
includes an inefficient check that a thread is free, when a request
completes.

A server application was taking 5% of its time running this check,
which is a .contains call on HashMap.values.

This CL replaces it with a use count on each IO thread, which is
checked against 0.

Bug: https://github.com/dart-lang/sdk/issues/50907
Change-Id: I6d13f9c62d36c34f165b2b4e32a48caeb6f9a77b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/278503
Reviewed-by: Zach Anderson <zra@google.com>
Commit-Queue: William Hesse <whesse@google.com>