Add information about how to write diagnostic documentation
I thought it would be good to capture some of the information we have
about how the documentation for the analyzer diagnostics are currently
written.

I'm mostly interested in making sure that the description is clear.

We can update this if there are changes we decide to make going forward.

Change-Id: I27c78ce04fb86d0263a1081d7c8f42c3ebfa7b0a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/278534
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Marya Belanger <mbelanger@google.com>