Rename TypeAnalyzer.getVariableType's parameter.
The previous name, `node`, was confusing, because the type is
`Variable`, not `Node`.  Although the concrete type used by the CFE
for `Variable` (`VariableDeclaration`) happens to be a type of `Node`,
the same is not true for the analyzer (which uses the type
`PromotableElement`).

Change-Id: Idf3a7c3101786f7d36fa9075053544307ba18519
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/278648
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>