[cfe] Insert synthetic break statements
With the patterns feature, switch cases without a break now implicitly
break instead of falling through to the next case (which was an error
anyway). We now insert a synthetic break statement in such cases.

To avoid crashes, the implementation to `TypeAnalyzerisAlwaysExhaustiveType` is also provided. This function
is not explicit test by this CL.

Closes #50624

Change-Id: I5d467072e26d79d0469e216bd21b460237af931b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/278746
Reviewed-by: Jens Johansen <jensj@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>