Flow analysis: fix synthetic break insertion for unreachable switch cases
Previously, if a switch case was unreachable, we would not insert a
synthetic break at the end of it.  Technically this is not a problem
(since a synthetic break is only required to prevent one case from
falling through to another at runtime, and unreachable code is by
definition never reached at runtime).  However, it makes it confusing
for CFE and back-end developers, and there's little harm in adding the
synthetic break anyway.

So with this change, we determine whether to add the synthetic break
by checking whether the bottom of the case block is reachable *from
its top* (rather than globally reachable).

Fixes #50994.

Bug: https://github.com/dart-lang/sdk/issues/50994
Change-Id: I17757b182c29da782457adc057b1b8a6fc91e55e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/278897
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>