[dart2wasm] Update Binaryen options for benchmarking
Remove the `--nominal` option, since this is obsolete and incompatible
with the singleton recursion groups we use for async and FFI support.

Add the `--type-ssa` and `--type-merging` optimization flags, which
improve the precision of Binaryen's type flow analysis and subsequent
devirtualization of method and closure calls.

Also fix a path error in the `compile_benchmark` script.

Change-Id: I8e9ac2278f9d68c9880760a0547b5911aee48a8e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/279015
Reviewed-by: Joshua Litt <joshualitt@google.com>
Commit-Queue: Aske Simon Christensen <askesc@google.com>