Add documentation for unnecessary_nan_comparison
I thought that this would be a nice small trial run for reviewing new
diagnostic documentation in a CL rather than in a separate doc.

It turns out that it does have one quirk, which is the need to update
the generator in order to add a link. Also note that this is the first
link to a site we don't control (the others are all to `dart.dev` or the
language repo on GitHub. I'm not sure whether that's something we want
to do.

Change-Id: I576a1a3a4f3daff938a4e93542f1cef9f2508a57
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/278640
Reviewed-by: Marya Belanger <mbelanger@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>