[gardening] Skip asset tests on RISC-V QEMU bot
The test tries to invoke itself with the Dart executable.
On QEMU, this is hard to find. (It would make more sense to compute
things on the host OS rather than doing self-invokes for QEMU, but
requires a significant refactoring of these tests.)

Test is already skipped on ARM QEMU.

Bug: https://github.com/dart-lang/sdk/issues/51067
Change-Id: I65d47141728223b835ac79d99c243baddb7ddcb0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/279346
Reviewed-by: Tess Strickland <sstrickl@google.com>
Commit-Queue: Daco Harkes <dacoharkes@google.com>